Scrolling moved to view
Moved scrolling to the view object, out of userInputComponent.'
Maks more sense thre. The userInputComponent adds it's owner to the view as a scroller.
Also, changed scrolling so it relies on how far past the border you are, not on how far you moved last. Greatly simplified things. No longer have to keep track of previous position!